+++ actions/host_delete.py
@@ -24,9 +24,7 @@ class HostDelete(ZabbixBaseAction):
         """
         self.connect()
 
-        zabbix_host = self.find_host(host)
-
-        host_id = zabbix_host['hostid']
+        host_id = self.find_host(host)
 
         try:
             self.client.host.delete(host_id)