+++ tests/test_host_delete.py
@@ -11,29 +11,26 @@ class HostDeleteTestCase(ZabbixBaseActionTestCase):
     __test__ = True
     action_cls = HostDelete
 
-    @mock.patch('lib.actions.ZabbixBaseAction.find_host')
     @mock.patch('lib.actions.ZabbixBaseAction.connect')
-    def test_run_connection_error(self, mock_connect, mock_find_host):
+    def test_run_connection_error(self, mock_connect):
         action = self.get_action_instance(self.full_config)
         mock_connect.side_effect = URLError('connection error')
         test_dict = {'host': "test"}
         host_dict = {'name': "test", 'hostid': '1'}
-        mock_find_host.return_vaue = host_dict
+        mock.MagicMock(return_value=host_dict['hostid'])
 
         with self.assertRaises(URLError):
             action.run(**test_dict)
 
-    @mock.patch('lib.actions.ZabbixBaseAction.find_host')
     @mock.patch('lib.actions.ZabbixBaseAction.connect')
-    def test_run_host_error(self, mock_connect, mock_find_host):
+    def test_run_host_error(self, mock_connect):
         action = self.get_action_instance(self.full_config)
         mock_connect.return_vaue = "connect return"
         test_dict = {'host': "test"}
         host_dict = {'name': "test", 'hostid': '1'}
-        mock_find_host.side_effect = ZabbixAPIException('host error')
-        mock_find_host.return_vaue = host_dict
+        action.find_host = mock.MagicMock(return_value=host_dict['hostid'],
+            side_effect=ZabbixAPIException('host error'))
         action.connect = mock_connect
-        action.find_host = mock_find_host
 
         with self.assertRaises(ZabbixAPIException):
             action.run(**test_dict)
@@ -46,7 +43,7 @@ class HostDeleteTestCase(ZabbixBaseActionTestCase):
         test_dict = {'host': "test"}
         host_dict = {'name': "test", 'hostid': '1'}
         action.connect = mock_connect
-        action.find_host = mock.MagicMock(return_value=host_dict)
+        action.find_host = mock.MagicMock(return_value=host_dict['hostid'])
         mock_client.host.delete.return_value = "delete return"
         action.client = mock_client
 
@@ -62,7 +59,7 @@ class HostDeleteTestCase(ZabbixBaseActionTestCase):
         test_dict = {'host': "test"}
         host_dict = {'name': "test", 'hostid': '1'}
         action.connect = mock_connect
-        action.find_host = mock.MagicMock(return_value=host_dict)
+        action.find_host = mock.MagicMock(return_value=host_dict['hostid'])
         mock_client.host.delete.side_effect = ZabbixAPIException('host error')
         mock_client.host.delete.return_value = "delete return"
         action.client = mock_client