+++ tests/test_maintenance_create_or_update.py
@@ -11,9 +11,8 @@ class MaintenanceCreateOrUpdateTestCase(ZabbixBaseActionTestCase):
     __test__ = True
     action_cls = MaintenanceCreateOrUpdate
 
-    @mock.patch('lib.actions.ZabbixBaseAction.find_host')
     @mock.patch('lib.actions.ZabbixBaseAction.connect')
-    def test_run_connection_error(self, mock_connect, mock_find_host):
+    def test_run_connection_error(self, mock_connect):
         action = self.get_action_instance(self.full_config)
         mock_connect.side_effect = URLError('connection error')
         test_dict = {'host': "test",
@@ -23,14 +22,13 @@ class MaintenanceCreateOrUpdateTestCase(ZabbixBaseActionTestCase):
                     'start_date': "2017-11-14 10:40",
                     'end_date': "2017-11-14 10:45"}
         host_dict = {'name': "test", 'hostid': '1'}
-        mock_find_host.return_vaue = host_dict
+        mock.MagicMock(return_value=host_dict['hostid'])
 
         with self.assertRaises(URLError):
             action.run(**test_dict)
 
-    @mock.patch('lib.actions.ZabbixBaseAction.find_host')
     @mock.patch('lib.actions.ZabbixBaseAction.connect')
-    def test_run_host_error(self, mock_connect, mock_find_host):
+    def test_run_host_error(self, mock_connect):
         action = self.get_action_instance(self.full_config)
         mock_connect.return_vaue = "connect return"
         test_dict = {'host': "test",
@@ -40,10 +38,9 @@ class MaintenanceCreateOrUpdateTestCase(ZabbixBaseActionTestCase):
                     'start_date': "2017-11-14 10:40",
                     'end_date': "2017-11-14 10:45"}
         host_dict = {'name': "test", 'hostid': '1'}
-        mock_find_host.side_effect = ZabbixAPIException('host error')
-        mock_find_host.return_vaue = host_dict
+        action.find_host = mock.MagicMock(return_value=host_dict['hostid'],
+            side_effect=ZabbixAPIException('host error'))
         action.connect = mock_connect
-        action.find_host = mock_find_host
 
         with self.assertRaises(ZabbixAPIException):
             action.run(**test_dict)
@@ -61,7 +58,7 @@ class MaintenanceCreateOrUpdateTestCase(ZabbixBaseActionTestCase):
         host_dict = {'name': "test", 'hostid': '1'}
         maintenance_dict = {'maintenanceids': ['1']}
         action.connect = mock_connect
-        action.find_host = mock.MagicMock(return_value=host_dict)
+        action.find_host = mock.MagicMock(return_value=host_dict['hostid'])
         action.maintenance_create_or_update = mock.MagicMock(return_value=maintenance_dict,
             side_effect=ZabbixAPIException('maintenance error'))
 
@@ -82,7 +79,7 @@ class MaintenanceCreateOrUpdateTestCase(ZabbixBaseActionTestCase):
         maintenance_dict = {'maintenanceids': ['1']}
         expected_return = {'maintenance_id': maintenance_dict['maintenanceids'][0]}
         action.connect = mock_connect
-        action.find_host = mock.MagicMock(return_value=host_dict)
+        action.find_host = mock.MagicMock(return_value=host_dict['hostid'])
         action.maintenance_create_or_update = mock.MagicMock(return_value=maintenance_dict)
 
         result = action.run(**test_dict)